requirements.txt
affine==2.4.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pyparsing==3.1.0
rasterio==1.3.7
requests==2.31.0
shapely==2.0.1
snuggs==1.4.7
urllib3==2.0.3
